feat(tab): get typings working with esm
This also makes it so interface augmentation works, which allows users to add properties to the props types through a declaration
